pyproject.toml
[project]
name = "djangoweb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.7",
    "django-allauth>=65.4.1",
    "django-htmx>=1.22.0",
    "django-redis>=5.4.0",
    "django-tables2>=2.7.5",
    "django-widget-tweaks>=1.5.0",
    "dotenv>=0.9.9",
    "gunicorn>=23.0.0",
    "mysqlclient>=2.2.7",
]

[dependency-groups]
dev = [
    "colorama>=0.4.6",
    "django-debug-toolbar>=5.0.1",
    "django-extensions>=3.2.3",
]
